+++ docs/changelog.md
@@ -39,4 +39,4 @@
 - rename enum variants with comment directive `@graphql({"mappings": "sql-value": "graphql_value""})`
 - bugfix: query with more than 50 fields fails
 - bugfix: @skip and @include directives missing from introspection schema
-- feature: Support for `and`, `OR` and `NOT` operators in filters
+- feature: Support for `and`, `or` and `NOT` operators in filters